+++ tests/testthat/test-sql.R
@@ -46,10 +46,15 @@ describe("glue_sql", {
     expect_identical(glue_sql("{var*}", .con = con, var = "a"), DBI::SQL("'a'"))
     expect_identical(glue_sql("{var*}", .con = con, var = letters[1:5]), DBI::SQL("'a', 'b', 'c', 'd', 'e'"))
   })
-  it('collapses values should return NULL for length zero vector', {
-    expect_identical(glue_sql("{var*}", .con = con, var = character()), DBI::SQL("NULL"))
-    expect_identical(glue_sql("{var*}", .con = con, var = DBI::SQL(character())), DBI::SQL("NULL"))
+  it('collapses empty values to empty string', {
+    expect_identical(glue_sql("{var*}", .con = con, var = NULL), DBI::SQL(""))
+    expect_identical(glue_sql("{var*}", .con = con, var = character()), DBI::SQL(""))
+    expect_identical(glue_sql("{var*}", .con = con, var = DBI::SQL(character())), DBI::SQL(""))
   })
+  it("mimics glue() when not collapsing", {
+    expect_equal(glue_sql("{var}", var = NULL), DBI::SQL(glue("{var}", var = NULL)))
+  })
+
   it("should return an SQL NULL by default for missing values", {
     var <- list(NA, NA_character_, NA_real_, NA_integer_)
     expect_identical(glue_sql("x = {var}", .con = con), rep(DBI::SQL("x = NULL"), 4))